Move AndroidLogcatFakeDeviceQuery to separate file
