+++ com.unity.mobile.android-logcat/Tests/Editor/AndroidLogcatTestDeviceQueryTests.cs
@@ -8,58 +8,6 @@ using Unity.Android.Logcat;
 using UnityEditor.Android;
 
 
-class AndroidLogcatFakeDeviceQuery : AndroidLogcatDeviceQueryBase
-{
-    List<string> m_QueuedInfos = new List<string>();
-
-
-    internal AndroidLogcatFakeDeviceQuery(IAndroidLogcatRuntime runtime)
-        : base(runtime)
-    {
-    }
-
-    internal void QueueDeviceInfos(string infos)
-    {
-        m_QueuedInfos.Add(infos);
-    }
-
-    internal override void UpdateConnectedDevicesList(bool synchronous)
-    {
-        m_Runtime.Dispatcher.Schedule(new AndroidLogcatRetrieveDeviceIdsInput() { adb = null, notifyListeners = true }, QueryDevicesAsync, IntegrateQueryDevices, synchronous);
-    }
-
-    private IAndroidLogcatTaskResult QueryDevicesAsync(IAndroidLogcatTaskInput input)
-    {
-        var result = new AndroidLogcatRetrieveDeviceIdsResult();
-        result.notifyListeners = ((AndroidLogcatRetrieveDeviceIdsInput)input).notifyListeners;
-
-        try
-        {
-            var adbOutput = m_QueuedInfos.Count > 0 ? m_QueuedInfos[0] : string.Empty;
-            if (m_QueuedInfos.Count > 0)
-                m_QueuedInfos.RemoveAt(0);
-
-            foreach (var line in adbOutput.Split(new[] { '\r', '\n' }, StringSplitOptions.RemoveEmptyEntries).Select(line => line.Trim()))
-            {
-                AndroidLogcatRetrieveDeviceIdsResult.DeviceInfo info;
-                if (ParseDeviceInfo(line, out info.id, out info.state))
-                    result.deviceInfo.Add(info);
-            }
-        }
-        catch (Exception)
-        {
-            result.deviceInfo = new List<AndroidLogcatRetrieveDeviceIdsResult.DeviceInfo>();
-        }
-
-        return result;
-    }
-
-    protected override IAndroidLogcatDevice CreateDevice(string deviceId)
-    {
-        return new AndroidLogcatFakeDevice90(deviceId);
-    }
-}
-
 class AndroidLogcatDeviceQueryTests : AndroidLogcatRuntimeTestBase
 {
     int m_UpdatedCount = 0;
